Modifiche Matteo 12:45 21 feb 24

+++ src/main/java/com/vino/vino/model/dao/WineDAO.java
@@ -17,7 +17,7 @@ public interface WineDAO {
             String vitigni,
             String provenance,
             String format,
-            int alcool,
+            Float alcool,
             String category,
             String description
     ) throws DuplicatedObjectException, DataTruncationException;
